+++ src/components/FilterBar.jsx
@@ -1,34 +1,15 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import {
-  Grid,
-  FormControl,
-  Select,
-  MenuItem,
-  Drawer,
-  useMediaQuery,
-  IconButton,
-} from '@mui/material';
-import AmountFilter from './AmountFilter';
-import GradeLevelFilter from './GradeLevelFilter';
-import qParams from '../lib/QueryParams';
+import { Grid, FormControl, Select, MenuItem, IconButton } from '@mui/material';
 import sortOptions, { DEADLINE_ASC } from '../lib/sortOptions';
-import MajorFilter from './MajorFilter';
 import TuneIcon from '@mui/icons-material/Tune';
 
-export default function FilterBar({ setQueryParam, queryParams }) {
-  const { minAmount, maxAmount, grades, majors } = queryParams;
-  const [drawerOpen, setDrawerOpen] = useState(false);
-
-  const isDesktop = useMediaQuery((theme) => theme.breakpoints.up('md'));
-
+export default function FilterBar({ setQueryParam, queryParams, openFn }) {
   return (
     <>
       <Grid container justifyContent="space-between" alignItems="center">
         <Grid item>
-          <IconButton
-            onClick={() => setDrawerOpen(true)}
-            sx={{ display: { sm: 'none' } }}>
+          <IconButton onClick={openFn} sx={{ display: { md: 'none' } }}>
             <TuneIcon />
           </IconButton>
         </Grid>
@@ -50,34 +31,6 @@ export default function FilterBar({ setQueryParam, queryParams }) {
           </FormControl>
         </Grid>
       </Grid>
-
-      <Drawer
-        variant={isDesktop ? 'permanent' : 'temporary'}
-        open={drawerOpen}
-        onClose={() => setDrawerOpen(!drawerOpen)}>
-        <MajorFilter
-          majors={majors}
-          onSelect={(m) => setQueryParam(qParams.MAJORS, m)}
-          onDelete={(m) =>
-            setQueryParam(
-              qParams.MAJORS,
-              majors.filter((major) => major !== m)
-            )
-          }
-        />
-
-        <GradeLevelFilter
-          grades={new Set(grades)}
-          changeFn={(e) => setQueryParam(qParams.GRADES, e)}
-        />
-
-        <AmountFilter
-          min={minAmount ?? 0}
-          max={maxAmount ?? 0}
-          onMinChange={(e) => setQueryParam(qParams.MIN_AMOUNT, e.target.value)}
-          onMaxChange={(e) => setQueryParam(qParams.MAX_AMOUNT, e.target.value)}
-        />
-      </Drawer>
     </>
   );
 }
@@ -89,6 +42,7 @@ FilterBar.propTypes = {
     maxAmount: PropTypes.number,
     sortBy: PropTypes.string,
   }),
+  openFn: PropTypes.func.isRequired,
 };
 FilterBar.defaultProps = {
   queryParams: {},
